requirements/common.txt
Django==1.11.2
django-cors-headers==2.1.0
django-crispy-forms==1.6.0
django-dbbackup==3.1.3
django-filter==1.0.4
django-haystack==2.6.0
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
Markdown==2.6.8
django-taggit==0.22.1
Whoosh==2.7.4